pyproject.toml
[tool.poetry]
name = "bnm-handler-by-ub-ln"
version = "0.1.0"
description = ""
authors = ['snorlax <snorlax@gmail.com>']
readme = "README.md"
packages = [{include = "bnm_handler_by_ub_ln"}]
source = []


[tool.poetry.dependencies]
python = "^3.8"
tomlkit = "^0.12.1"
python-dotenv = "0.21.1"
pytz = "*"
casestyle = "^0.0.4"
pydantic-settings = "^2.4.0"           
binance-connector = "^3.7.0"

[tool.poetry.scripts]
patp = "scripts:patp"
minp = "scripts:minp"
majp = "scripts:majp"
start = "bnm_handler_by_ub_ln.__main__:main"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pylint = "^2.17.5"
pre-commit = "^3.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"
mypy = "^1.5.1"
types-pytz = "*"
parver = "^0.5"


[tool.current-project]
poetry_publish_enabled = true
docker_enabled = false
workflow_enabled = false


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
